+++ src/mdast-process-images.js
@@ -21,18 +21,42 @@ export class TooManyImagesError extends Error {
  * @param {object} tree
  * @param {MediaHandler} mediaHandler
  * @param {string} baseUrl
+ * @param {Array<string>} externalImageUrlPrefixes Array of url prefixes to detect external images
  */
-export async function processImages(log, tree, mediaHandler, baseUrl) {
+export async function processImages(
+  log,
+  tree,
+  mediaHandler,
+  baseUrl,
+  externalImageUrlPrefixes = [],
+  maxImages = 200,
+) {
   if (!mediaHandler) {
     return;
   }
   // gather all image nodes
   const images = new Map();
+  // Convert externalImageUrlPrefixes to an array if not already
+  if (!Array.isArray(externalImageUrlPrefixes)) {
+    // eslint-disable-next-line no-param-reassign
+    externalImageUrlPrefixes = [externalImageUrlPrefixes];
+  }
+
   const register = (node) => {
-    if (images.has(node.url)) {
-      images.get(node.url).push(node);
+    // Check if this is an external image
+    const { url = '' } = node;
+    const isExternalImage = externalImageUrlPrefixes
+      .some((externalImageUrlPrefix) => url.startsWith(externalImageUrlPrefix));
+    if (isExternalImage) {
+      log.debug(`Skipping upload for external image: ${url}`);
+      return;
+    }
+
+    // Regular image processing
+    if (images.has(url)) {
+      images.get(url).push(node);
     } else {
-      images.set(node.url, [node]);
+      images.set(url, [node]);
     }
   };
 
@@ -50,11 +74,11 @@ export async function processImages(log, tree, mediaHandler, baseUrl) {
     return CONTINUE;
   });
 
-  if (images.size > 200) {
-    throw new TooManyImagesError(`maximum number of images reached: ${images.size} of 200 max.`);
+  if (images.size > maxImages) {
+    throw new TooManyImagesError(`maximum number of images reached: ${images.size} of ${maxImages} max.`);
   }
 
-  // upload images
+  // upload regular images
   await processQueue(images.entries(), async ([url, nodes]) => {
     try {
       const blob = await mediaHandler.getBlob(url, baseUrl);